Added Tile Button Class and Created Game Field
Created an alternative version of the Tile class called TileButton, where it just extends from Button instead of DrawNode. Modified add tile function in board to compensate for change. Also could not get the Board Manager to work for me, so I created board through a function in the main game layer instead for testing purposes.